backend/requirements.txt
web3==6.10.0
python-dotenv==1.0.1
